+++ app/src/main/java/com/example/overtime/ui/component/ZetaAlertDialog.kt
@@ -1,0 +1,38 @@
+package com.example.overtime.ui.component
+
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.style.TextAlign
+
+@Composable
+fun ZetaAlertDialog(
+    title: String,
+    message: String,
+    confirmText: String,
+    onConfirmClick: () -> Unit,
+    onDismissClick: () -> Unit
+) {
+    val scroll = rememberScrollState(0)
+
+    AlertDialog(
+        onDismissRequest = { onDismissClick() },
+        title = { Text(text = title) },
+        text = {
+            Text(
+                text = message,
+                textAlign = TextAlign.Justify,
+                modifier = Modifier.verticalScroll(scroll)
+            )
+        },
+        confirmButton = {
+            Button(onClick = { onConfirmClick() }) {
+                Text(text = confirmText)
+            }
+        }
+    )
+}
