+++ app/src/main/java/com/example/overtime/ui/component/ZetaSpaceHeight.kt
@@ -1,0 +1,16 @@
+package com.example.overtime.ui.component
+
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+
+
+@Composable
+fun ZetaSpaceHeight(size: Dp = 10.dp) {
+    Spacer(modifier = Modifier.height(size))
+}
+
+
